Progress in refactoring raytracing
Moved some logic into new files, continuing to rework old interfaces to
connect the front to back logic. Should compile now.
